requirements.txt
Numpy
scikit_image
scikit-learn==0.22.2.post1
opencv_python
tensorflow
requests
matplotlib
pandas==0.23.3
Keras_Applications==1.0.8
joblib==0.14.1
tqdm==4.44.1
PySimpleGUI==4.28.0
